Update build and export to match new plugin model
